grai-integrations/source-bigquery/pyproject.toml
[tool.poetry]
name = "grai_source_bigquery"
version = "0.0.1"
description = ""
authors = ["Edward Louth <edward@grai.io>"]
license = "ELv2"
packages = [
    { include = "grai_source_bigquery", from = "src" },
]
[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.9.1"
grai-client = "^0.2.0"
PyYAML = "^6.0"
multimethod = "^1.8"
grai-schemas = "^0.1.9"
google-cloud-bigquery = "^3.5.0"

[tool.poetry.group.dev.dependencies]
black = "^22.6.0"
mypy = "^0.971"
isort = "^5.10.1"
types-PyYAML = "^6.0.11"
pytest = "^7.2.0"
python-dotenv = "^0.21.1"

[tool.isort]
profile = "black"
known_first_party = "grai_source_bigquery"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
